+++ packages/gatsby-transformer-sharp/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.6.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@3.5.0-next.0...gatsby-transformer-sharp@3.6.0-next.0) (2021-05-06)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
 # [3.5.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@3.4.0-next.0...gatsby-transformer-sharp@3.5.0-next.0) (2021-04-23)
 
 **Note:** Version bump only for package gatsby-transformer-sharp